+++ packages/react/ds/src/header/types.ts
@@ -20,6 +20,7 @@ type ConditionalProps =
   | {
       label?: string;
       ariaLabel?: string;
+      closeLabel?: string;
       icon?: IconId;
       component: React.ReactNode;
       href?: never;
@@ -32,6 +33,7 @@ type ConditionalProps =
   | {
       label?: never;
       ariaLabel?: string;
+      closeLabel?: string;
       icon?: never;
       component?: never;
       href?: never;
@@ -44,6 +46,7 @@ type ConditionalProps =
   | {
       label?: string;
       ariaLabel?: string;
+      closeLabel?: string;
       icon?: IconId;
       component?: never;
       href?: string;
@@ -56,6 +59,7 @@ type ConditionalProps =
   | {
       label?: never;
       ariaLabel?: string;
+      closeLabel?: string;
       icon?: never;
       component: React.ReactNode;
       href?: never;